refactor: 構造を見直し
