+++ src/components/UI/DropdownMenu/index.tsx
@@ -2,12 +2,12 @@
 
 import { useButton } from '@react-aria/button';
 import { useInteractOutside } from '@react-aria/interactions';
-import { useMenu, useMenuItem } from '@react-aria/menu';
-import type { ReactNode, RefObject } from 'react';
-import { Children, isValidElement, useId, useMemo, useRef } from 'react';
-import type { Node } from 'react-stately';
-import { Item, useMenuTriggerState, useTreeState } from 'react-stately';
-import { styled } from '@/ui/styled';
+import type { ReactNode } from 'react';
+import { useId, useRef } from 'react';
+import { useMenuTriggerState } from 'react-stately';
+
+import { MenuPopover } from './MenuPopover';
+import { Container, Trigger } from './styles';
 
 type MenuPosition = 'left' | 'right';
 
@@ -64,171 +64,3 @@ export function DropdownMenu({ title, children, menuHorizontalPosition = 'right'
     </Container>
   );
 }
-
-type MenuPopoverProps = {
-  children: ReactNode;
-  onClose: () => void;
-  autoFocus?: boolean | 'first' | 'last';
-  menuRef: RefObject<HTMLDivElement>;
-  isOpen: boolean;
-  'data-position': MenuPosition;
-  'aria-labelledby': string;
-};
-
-/**
- * ReactNodeからテキストコンテンツを再帰的に抽出
- * React AriaのtextValue要求を満たすために使用（スクリーンリーダー対応）
- */
-function extractTextContent(node: ReactNode): string {
-  if (typeof node === 'string' || typeof node === 'number') {
-    return String(node);
-  }
-
-  if (Array.isArray(node)) {
-    return node.map(extractTextContent).join('');
-  }
-
-  if (isValidElement(node)) {
-    const { children } = node.props as { children?: ReactNode };
-    return children ? extractTextContent(children) : '';
-  }
-  return '';
-}
-
-function MenuPopover({
-  children,
-  onClose,
-  autoFocus,
-  menuRef,
-  isOpen,
-  'data-position': dataPosition,
-  'aria-labelledby': ariaLabelledby,
-}: MenuPopoverProps) {
-  // 子要素を React Stately の Item にラップ（childrenが変更時のみ再計算）
-  const menuItems = useMemo(() => {
-    return Children.toArray(children).map((child, index) => {
-      const textValue = extractTextContent(child);
-      // 既存のkeyを優先、なければtextValueとindexから生成
-      const itemKey = isValidElement(child) && child.key ? child.key : `menu-item-${textValue}-${index}`;
-      return (
-        <Item key={itemKey} textValue={textValue}>
-          {child}
-        </Item>
-      );
-    });
-  }, [children]);
-
-  // メニュー選択を無効化（ナビゲーション目的のためselectionMode: 'none'）
-  const state = useTreeState({ selectionMode: 'none', children: menuItems });
-
-  const { menuProps } = useMenu({ autoFocus, onClose, 'aria-labelledby': ariaLabelledby }, state, menuRef);
-
-  return (
-    <Content {...menuProps} aria-expanded={isOpen} data-position={dataPosition} ref={menuRef} role="menu">
-      {Array.from(state.collection).map((item) => (
-        <MenuItem item={item} key={item.key} onAction={onClose} state={state} />
-      ))}
-    </Content>
-  );
-}
-
-type MenuItemProps = {
-  item: Node<object>;
-  state: ReturnType<typeof useTreeState>;
-  onAction: () => void;
-};
-
-function MenuItem({ item, state, onAction }: MenuItemProps) {
-  const ref = useRef<HTMLDivElement>(null);
-  const { menuItemProps } = useMenuItem(
-    {
-      key: item.key,
-      onAction,
-    },
-    state,
-    ref,
-  );
-
-  return (
-    <MenuItemContainer {...menuItemProps} ref={ref}>
-      {item.rendered}
-    </MenuItemContainer>
-  );
-}
-
-export const Container = styled.div`
-  position: relative;
-  display: flex;
-`;
-
-export const Trigger = styled.button`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: calc(var(--sizes-icon-sm) * 2);
-  height: calc(var(--sizes-icon-sm) * 2);
-
-  &:has(+ [aria-expanded='true']) {
-    &::after {
-      background-color: var(--colors-gray-a-200);
-      opacity: 1;
-      transform: scale(1);
-    }
-  }
-`;
-
-export const Content = styled.div`
-  position: absolute;
-  top: 100%;
-  z-index: var(--z-index-base);
-  visibility: hidden;
-  min-width: max-content;
-  height: fit-content;
-  padding: var(--spacing-½);
-  pointer-events: none;
-  background-color: var(--colors-white);
-  border: var(--border-widths-thin) solid var(--colors-gray-a-200);
-  border-radius: var(--radii-sm);
-  box-shadow: var(--shadows-md);
-  opacity: 0;
-  transform: scale(0.95);
-  transition:
-    opacity var(--transition-normal),
-    transform var(--transition-normal),
-    visibility 0s var(--durations-normal);
-
-  &[data-position='left'] {
-    left: 0;
-    transform-origin: 0 0;
-  }
-
-  &[data-position='right'] {
-    right: 0;
-    transform-origin: 100% 0;
-  }
-
-  &[aria-expanded='true'] {
-    visibility: visible;
-    pointer-events: auto;
-    transition-delay: 0s;
-    animation: dropdownEnter var(--transition-slow) forwards;
-  }
-
-  @media (prefers-reduced-motion: reduce) {
-    &[aria-expanded='true'] {
-      opacity: 1;
-      transform: scale(1);
-      animation: none;
-    }
-  }
-`;
-
-export const MenuItemContainer = styled.div`
-  & > a {
-    display: flex;
-    align-items: center;
-    padding: var(--spacing-½) var(--spacing-1);
-    line-height: var(--line-heights-lg);
-    border-radius: var(--radii-sm);
-  }
-`;
